Add RS and ORS variables for changing input/output row separators
This CL adds a first sketch of rxLineSeparator which is capable
of splitting io.Reader into chunks (lines) separated by the given
regexp.

Also a bug concerning the retrieval of the Filename was fixed, although
not tested. Reading of a Source is wrapped by a buffer. At some point
the Filename method would return the name of one of the next sources
although the buffer might contain bytes from the previous source.